+++ core-tests/jvm/src/test/scala/zio/TestRuntime.scala
@@ -6,7 +6,9 @@ import org.specs2.matcher.Expectations
 import org.specs2.matcher.TerminationMatchers.terminate
 import org.specs2.specification.{ Around, AroundEach, AroundTimeout }
 
-import scala.concurrent.duration._
+import scala.concurrent.duration.{ Duration => SDuration, MICROSECONDS }
+
+import zio.duration._
 
 abstract class TestRuntime(implicit ee: org.specs2.concurrent.ExecutionEnv)
     extends BaseCrossPlatformSpec
@@ -19,15 +21,26 @@ abstract class TestRuntime(implicit ee: org.specs2.concurrent.ExecutionEnv)
       case other                                 => other
     }
 
-  override final def aroundTimeout(to: Duration)(implicit ee: ExecutionEnv): Around =
+  override final def aroundTimeout(to: SDuration)(implicit ee: ExecutionEnv): Around =
     new Around {
       def around[T: AsResult](t: => T): Result = {
         lazy val result = t
-        val termination = terminate(retries = 1000, sleep = (to.toMicros / 1000).micros)
+        val termination = terminate(retries = 1000, sleep = SDuration(to.toMicros / 1000, MICROSECONDS))
           .orSkip(_ => "TIMEOUT: " + to)(Expectations.createExpectable(result))
 
         if (!termination.toResult.isSkipped) AsResult(result)
         else termination.toResult
       }
+    }
+
+  final def flaky(
+    v: => ZIO[Environment, Any, org.specs2.matcher.MatchResult[Any]]
+  ): org.specs2.matcher.MatchResult[Any] =
+    eventually(unsafeRun(v.timeout(1.second)).get)
+
+  def nonFlaky(v: => ZIO[Environment, Any, org.specs2.matcher.MatchResult[Any]]): org.specs2.matcher.MatchResult[Any] =
+    (1 to 100).foldLeft[org.specs2.matcher.MatchResult[Any]](true must_=== true) {
+      case (acc, _) =>
+        acc and unsafeRun(v)
     }
 }
